requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
slack_sdk==3.33.4
